# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truefoods Sanifu API",
        "description": "API for Truefoods Sanifu integration with NetSuite. This API provides endpoints for managing items, customers, customer purchases, and sales orders.",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://netsuiteintegrator.dynamicsserv.com:4140",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/truefoods-sanifu/get/items": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Get Items",
                "description": "Retrieve a paginated list of items with optional filters",
                "operationId": "ad2ecaddc98c7a97d3dac7e252f61906",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "(mandatory) Company identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 6
                    },
                    {
                        "name": "environment",
                        "in": "query",
                        "description": "(mandatory) Environment type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sandbox",
                                "production"
                            ]
                        },
                        "example": "sandbox"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "example": 1
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        },
                        "example": 50
                    },
                    {
                        "name": "itemId",
                        "in": "query",
                        "description": "Filter by specific item ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "displayName",
                        "in": "query",
                        "description": "Filter by item display name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by item category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isInactive",
                        "in": "query",
                        "description": "Include inactive items",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "includeLocations",
                        "in": "query",
                        "description": "Include location data in response",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "includePricing",
                        "in": "query",
                        "description": "Include pricing data in response",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}